Add more condition for cancel enrollment

+++ src/enrollment/services/Enrollment.service.ts
@@ -68,14 +68,18 @@ export class EnrollmentService {
     }
 
     async removeEnrolledCourse(classId: number, userId: string): Promise<void>{
-        const classData = await this.classRepository.findOne({where: {id: classId}});
+        const classData = await this.classRepository.findOne({where: {id: classId}, relations: ['semester']});
         if(!classData) throw new BadRequestException('Không tìm thấy lớp học. Vui lòng kiểm tra lại!');
         if(classData.status === ClassStatus.FINISHED) throw new BadRequestException('Lớp học đã được bắt đầu! Không thể hủy đăng ký!');
-        
+
+        const now = new Date();
+        if(now.getTime() < new Date(classData.semester.startRegistration).getTime() || now.getTime() > new Date(classData.semester.endDateRegistration).getTime()) throw new BadRequestException('Không thể hủy đăng ký lớp học này!');
+
         const studentEntity = new Student();
         studentEntity.id = userId;
         const enrollment = await this.enrollmentRepository.findOne({where: {student: studentEntity, class: classData}});
         if(!enrollment) throw new BadRequestException('Bạn chưa đăng ký lớp học này!');
+        if(enrollment.isEnrolled) throw new BadRequestException('Không thể hủy đăng ký lớp học này!');
         await this.enrollmentRepository.remove(enrollment);
         classData.currentStudents -= 1;
         await this.classRepository.save(classData);
